fix(quotes): wire price + effectiveDate into create shorthand
After #312 merged, AddQuoteLineItemInput requires `price` and
`effectiveDate`. The two-call shorthand in #311's quotes create
broke at the type level until those fields were threaded through.

Extracted `resolveListPrice` and `resolveEffectiveDate` from
#312's line-items add command into a shared
`packages/cli/src/lib/quote-line-item-defaults.ts` so both call
sites use the same defaults. Same suggestedRetailPrice + today-UTC
behavior as #312.

When no list price is available for the chosen billing term, the
shorthand falls back with a clear error pointing at the explicit
two-step `quotes create` + `quotes line-items add --price ...`
path. Partners with custom pricing follow that path.
